+++ workon/git.py
@@ -9,7 +9,7 @@ class GITError(Exception):
 
 def get_stash_info(directory):
     """Return stash info under `directory`."""
-    logging.info('Checking GIT stashes under "%s"', directory)
+    logging.debug('Checking GIT stashes under "%s"', directory)
     return subprocess.run(
         'git stash list'.split(), cwd=directory,
         capture_output=True, text=True, check=False
@@ -21,7 +21,7 @@ def get_unpushed_branches_info(directory) -> str:
 
     Format is: <commit> (<branch>) <commit_message>
     """
-    logging.info('Checking for unpushed GIT commits under "%s"', directory)
+    logging.debug('Checking for unpushed GIT commits under "%s"', directory)
     return subprocess.run(
         'git log --branches --not --remotes --decorate --oneline'.split(),
         cwd=directory, capture_output=True, text=True, check=False
@@ -30,7 +30,7 @@ def get_unpushed_branches_info(directory) -> str:
 
 def get_unstaged_info(directory) -> str:
     """Return information about unstaged changes."""
-    logging.info('Checking for unstaged changes under "%s"', directory)
+    logging.debug('Checking for unstaged changes under "%s"', directory)
     return subprocess.run(
         'git status --short'.split(),
         cwd=directory, capture_output=True, text=True, check=False
@@ -40,7 +40,7 @@ def get_unstaged_info(directory) -> str:
 def clone(source, destination):
     """Clone a project from GIT `source` to `destination` directory."""
     try:
-        logging.info('Cloning "%s" to "%s"', source, destination)
+        logging.debug('Cloning "%s" to "%s"', source, destination)
         subprocess.run(
             ['git', 'clone', source, destination],
             check=True, capture_output=True, text=True