+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * `start` command now opens a project if it exists
+* Default logging level is INFO (was ERROR before)
 
 
 ## [0.2.1] - 2021-06-23